move declarative_argparse to common.
