+++ harness/determined/cli/shell.py
@@ -12,6 +12,7 @@ from determined.common import api
 from determined.common.api import request
 from determined.common.api.authentication import authentication_required
 from determined.common.check import check_eq, check_len
+from determined.common.declarative_argparse import Arg, Cmd
 
 from . import render
 from .command import (
@@ -25,7 +26,6 @@ from .command import (
     parse_config,
     render_event_stream,
 )
-from .declarative_argparse import Arg, Cmd
 
 
 @authentication_required
